+++ inst/sql/insert_stem_table_from_fluidin.sql
@@ -94,7 +94,8 @@ SELECT
     visit_occurrence_id,
     NULL AS visit_detail_id,
 
-    LEFT(CONCAT('Fluid In: ', s.source_value), 50) AS source_value,
+    -- OMOP CDM compliant: LEFT(CONCAT('Fluid In: ', s.source_value), 50) AS source_value,
+    LEFT(CONCAT('Fluid In: ', s.source_value), 255) AS source_value,
     NULL AS source_concept_id,
 
     -- stem_table for drug_exposure records (temporarily) stores fluidin in